requirements.dev.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
black==23.1.0
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.1.0
flake8==6.0.0
frozenlist==1.3.3
greenlet==1.1.3.post0
idna==3.4
iniconfig==2.0.0
marshmallow==3.18.0
marshmallow-dataclass==8.5.9
marshmallow-enum==1.5.1
mccabe==0.7.0
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pluggy==1.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyHamcrest==2.0.4
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.20.3
pytest-mock==3.10.0
pytz==2022.7.1
PyYAML==6.0
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.39.0
tomli==2.0.1
typing-inspect==0.8.0
typing_extensions==4.4.0
yarl==1.8.2
